Fixed MimeReader to support MimePart content larger than 2GB
Fixes issue #1252

+++ MimeKit/AsyncMimeReader.cs
@@ -366,7 +366,7 @@ namespace MimeKit {
 		{
 			int maxBoundaryLength = Math.Max (ReadAheadSize, GetMaxBoundaryLength ());
 			var formats = new bool[2];
-			int contentLength = 0;
+			long contentLength = 0;
 			bool incomplete = false;
 			bool midline = false;
